requirements.txt
tensorflow-cpu
numpy
librosa
scikit-learn
joblib
scikit-image